+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 5.7
 
+ - Expanded EquationTransform support.
+
 # 5.6
 
  - Fix inconsistent position of Circle effect when using updateLocations: false